requirements.txt
asgiref==3.9.1
Django==5.2.4
django-filter==25.1
djangorestframework==3.16.0
djangorestframework-jsonapi==8.0.0
drf-nested-routers==0.94.2
inflection==0.5.1
psycopg2-binary==2.9.10
sqlparse==0.5.3
